Add localization support for prayer-related texts and improve UI components
- Updated English, Persian, French, and Turkish language files to include new keys for prayer conditions, compass, and location messages.
- Refactored the IbadetSayfa and WidgetAyarlariSayfa to utilize localized strings for prayer conditions and widget previews.
- Enhanced AuroraSayacWidget and MinimalSayacWidget to display prayer times using localized strings.
- Improved code formatting and readability across various files.

+++ android/app/src/main/kotlin/com/example/huzur_vakti/alarm/BootReceiver.kt
@@ -101,7 +101,7 @@ class BootReceiver : BroadcastReceiver() {
                     "gunes" -> 45
                     else -> 15
                 }
-                val erkenDakika = prefs.getInt("flutter.erken_$vakitKey", varsayilanErken)
+                val erkenDakika = prefs.getLong("flutter.erken_$vakitKey", varsayilanErken.toLong()).toInt()
                 
                 // Ses dosyası
                 val sesDosyasi = prefs.getString("flutter.bildirim_sesi_$vakitKey", "ding_dong.mp3") ?: "ding_dong.mp3"
